refactor: split vertex-log.ts by buildctl-subcommand parsing stage, move host aggregation into report/build/
