+++ core/lib/log/vertex-log-parser.test.ts
@@ -3,12 +3,7 @@ import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
 import { dirname, join } from "node:path";
 import { fileURLToPath } from "node:url";
-import {
-  selectAllRefs,
-  parseAllowedRequestsFromText,
-  parseVertexAllowedLog,
-  aggregateAllowedHosts,
-} from "./vertex-log.ts";
+import { parseVertexAllowedLog } from "./vertex-log-parser.ts";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const fixturesDir = join(__dirname, "__fixtures__");
@@ -18,7 +13,6 @@ const fixturesDir = join(__dirname, "__fixtures__");
 // allowed RUN; a final stage COPYing from stage2). stage1/stage2's RUN
 // vertices have overlapping `started` timestamps, since independent stages
 // can run concurrently.
-const MULTISTAGE_HISTORIES = readFileSync(join(fixturesDir, "histories.json"), "utf8");
 const MULTISTAGE_RAWJSON = readFileSync(join(fixturesDir, "multistage-rawjson.json"), "utf8");
 
 // Captured from a live moby/buildkit v0.31.1 explicit-mode container building
@@ -31,117 +25,6 @@ const PADDED_STEPS_RAWJSON = readFileSync(join(fixturesDir, "padded-steps-rawjso
 function encode(text: string) {
   return Buffer.from(text, "utf8").toString("base64");
 }
-
-describe("selectAllRefs", () => {
-  it("returns the ref from a real single-record histories.json capture", () => {
-    assert.deepEqual(selectAllRefs(MULTISTAGE_HISTORIES), ["jsvnbiyvnlslxxui8ap3i2na2"]);
-  });
-
-  it("returns every ref ordered oldest-first by CreatedAt", () => {
-    const historiesText = [
-      JSON.stringify({ type: 1, record: { Ref: "newer", CreatedAt: { seconds: 200, nanos: 0 } } }),
-      JSON.stringify({ type: 1, record: { Ref: "older", CreatedAt: { seconds: 100, nanos: 0 } } }),
-    ].join("\n");
-    assert.deepEqual(selectAllRefs(historiesText), ["older", "newer"]);
-  });
-
-  it("breaks ties within the same second using nanos", () => {
-    const historiesText = [
-      JSON.stringify({
-        type: 1,
-        record: { Ref: "later-nanos", CreatedAt: { seconds: 100, nanos: 999 } },
-      }),
-      JSON.stringify({
-        type: 1,
-        record: { Ref: "earlier-nanos", CreatedAt: { seconds: 100, nanos: 500 } },
-      }),
-    ].join("\n");
-    assert.deepEqual(selectAllRefs(historiesText), ["earlier-nanos", "later-nanos"]);
-  });
-
-  it("deduplicates a ref reported on multiple lines as its build progresses", () => {
-    const historiesText = [
-      JSON.stringify({
-        type: 1,
-        record: { Ref: "build-1", CreatedAt: { seconds: 100, nanos: 0 } },
-      }),
-      JSON.stringify({
-        type: 1,
-        record: {
-          Ref: "build-1",
-          CreatedAt: { seconds: 100, nanos: 0 },
-          CompletedAt: { seconds: 101, nanos: 0 },
-        },
-      }),
-    ].join("\n");
-    assert.deepEqual(selectAllRefs(historiesText), ["build-1"]);
-  });
-
-  it("ignores blank lines and records with no Ref/CreatedAt", () => {
-    const historiesText = [
-      "",
-      JSON.stringify({ type: 1, record: {} }),
-      "",
-      MULTISTAGE_HISTORIES.trim(),
-    ].join("\n");
-    assert.deepEqual(selectAllRefs(historiesText), ["jsvnbiyvnlslxxui8ap3i2na2"]);
-  });
-
-  it("returns an empty array when there are no records at all", () => {
-    assert.deepEqual(selectAllRefs(""), []);
-  });
-});
-
-describe("parseAllowedRequestsFromText", () => {
-  it("parses a real 'proxy network requests:' block with a status code", () => {
-    const text = ["proxy network requests:", "- GET https://allowed.example.com/ -> 200"].join(
-      "\n",
-    );
-    assert.deepEqual(parseAllowedRequestsFromText(text), [
-      { method: "GET", url: "https://allowed.example.com/", status: 200 },
-    ]);
-  });
-
-  it("parses a request line with no status code at all", () => {
-    const text = ["proxy network requests:", "- GET https://allowed.example.com/"].join("\n");
-    assert.deepEqual(parseAllowedRequestsFromText(text), [
-      { method: "GET", url: "https://allowed.example.com/" },
-    ]);
-  });
-
-  it("parses multiple entries under one block, stopping at the next unrelated line", () => {
-    const text = [
-      "proxy network requests:",
-      "- GET https://allowed.example.com/ -> 200",
-      "- GET https://sub.wildcard.example.com/ -> 200",
-      'time="x" level=debug msg="Evaluated source policy" error="denied"',
-    ].join("\n");
-    assert.deepEqual(parseAllowedRequestsFromText(text), [
-      { method: "GET", url: "https://allowed.example.com/", status: 200 },
-      { method: "GET", url: "https://sub.wildcard.example.com/", status: 200 },
-    ]);
-  });
-
-  it("parses multiple separate blocks", () => {
-    const text = [
-      "proxy network requests:",
-      "- GET https://allowed.example.com/ -> 200",
-      'time="x" level=debug msg="> creating abc"',
-      "proxy network requests:",
-      "- GET http://allowed.example.com:8080/ -> 200",
-    ].join("\n");
-    assert.equal(parseAllowedRequestsFromText(text).length, 2);
-  });
-
-  it("does not misinterpret unrelated '- ' output as a request line outside a block", () => {
-    const text = "- rw-r--r-- 1 root root 0 file.txt";
-    assert.deepEqual(parseAllowedRequestsFromText(text), []);
-  });
-
-  it("returns an empty array when no block is present", () => {
-    assert.deepEqual(parseAllowedRequestsFromText('time="x" level=info msg="found worker"'), []);
-  });
-});
 
 describe("parseVertexAllowedLog", () => {
   it("captures every single-digit step despite BuildKit's leading-space padding (regression: a real 15-step build)", () => {
@@ -388,58 +271,5 @@ describe("parseVertexAllowedLog", () => {
     const result = parseVertexAllowedLog(`${doc1}\nnot json at all\n`);
     assert.equal(result.length, 1);
     assert.equal(result[0].entries[0].url, "https://one.example.com/");
-  });
-});
-
-describe("aggregateAllowedHosts", () => {
-  it("aggregates entries across multiple vertices within one build", () => {
-    const builds = [
-      [
-        { entries: [{ method: "GET", url: "https://allowed.example.com/", status: 200 }] },
-        { entries: [{ method: "GET", url: "https://allowed.example.com/", status: 200 }] },
-      ],
-    ];
-    assert.deepEqual(aggregateAllowedHosts(builds, "ALLOWED"), [
-      { host: "allowed.example.com", port: "443", ruleType: "HTTPS", reason: "-", count: 2 },
-    ]);
-  });
-
-  it("aggregates entries across multiple builds", () => {
-    const builds = [
-      [{ entries: [{ method: "GET", url: "https://allowed.example.com/one" }] }],
-      [{ entries: [{ method: "GET", url: "https://allowed.example.com/two" }] }],
-    ];
-    const result = aggregateAllowedHosts(builds, "ALLOWED");
-    assert.equal(result.length, 1);
-    assert.equal(result[0].count, 2);
-  });
-
-  it("uses the given decision label", () => {
-    const builds = [[{ entries: [{ method: "GET", url: "https://allowed.example.com/" }] }]];
-    assert.equal(aggregateAllowedHosts(builds, "ALLOWED")[0]?.count, 1);
-    // decision itself isn't part of the aggregated shape (aggregate() drops it),
-    // but distinct decisions must not collide during aggregation
-    const mixed = [[{ entries: [{ method: "GET", url: "https://allowed.example.com/" }] }]];
-    assert.deepEqual(aggregateAllowedHosts(mixed, "AUDIT"), [
-      { host: "allowed.example.com", port: "443", ruleType: "HTTPS", reason: "-", count: 1 },
-    ]);
-  });
-
-  it("skips vertices with no entries", () => {
-    const builds = [[{ entries: [] }]];
-    assert.deepEqual(aggregateAllowedHosts(builds, "ALLOWED"), []);
-  });
-
-  it("returns an empty array for no builds at all", () => {
-    assert.deepEqual(aggregateAllowedHosts([], "ALLOWED"), []);
-  });
-
-  it("resolves host/port the same way as core/lib/log/parse-identifier.ts's parseIdentifier", () => {
-    const builds = [
-      [{ entries: [{ method: "GET", url: "http://allowed.example.com:8080/path" }] }],
-    ];
-    assert.deepEqual(aggregateAllowedHosts(builds, "ALLOWED"), [
-      { host: "allowed.example.com", port: "8080", ruleType: "HTTP", reason: "-", count: 1 },
-    ]);
   });
 });
